Refactor UserPreferences to move the validation code to a separate module, concerns/parameter_validations.rb. Updated validation to expect an ActionController::Parameters object.
It was hard to use the VALIDATIONS structure to automatically
permit parameters using ActionController::Parameters#permit, so I punted
and just call permit! on the paramters. This is safe nonetheless,
because only validated parameters whose key is in VALIDATIONS can be
included in the scrubbed values for an instantiated UserPreferences
object.