Refactoring to provide a clearer link between statistics and indicators
I realised that the association between statistics and indicators was
not very robust. This refactoring aims to correct that. In particular,
we have a set of volume data (all property types, property status
and funding status) that doesn't fit very well as a set of standalone
graphs. However, to integrate volume into the other themes means
being to say when an indicator is useful (the volume indicator is not
useful in the context of buyer status, since there's no data collected),
and also which statistics in a theme can have volume (e.g in the
property type theme, only the 'all property types' statistic has
volume). Altogether, this means that the previous approach of two static
sets of statistics, those with indicators and those without, no longer
works.

The basis of the refactoring is to promote UkhpiTheme, UkhpiStatistic
and UkhpiIndicator up to being full classes instead of Structs. Then we
can add the additional logic into those classes: e.g. is the volume
indicator appropriate to this theme? Then we can embed the results of
that computation into the JSON serialisation that is passed initially to
JavaScript via the `data-` attributes on DataViews. DataView now
computes an `availableStatistics` property, rather than relying on the
fixed static list. The distinction between qualified and non-qualified
views can be removed.

I also took the opportunity to clean up some irritations, including
changing the theme names to reflect HMLR terminology, and fixing some
JavaScript identifiers to use camel-case rather than snake-case.

This refactoring also makes it easier to build a UI for the location
comparison page.