Further updates to improve the location search dialogue.
Ensure that the link between search selections and map selections
is made by auto-selecting the map layer and map region when a
search result is clicked on.

Use a popover for the search results list, because otherwise the dialog
box grows in height and is likely to overflow on smaller screens.