Fix for an issue in which a once-opened zoom graph would not properly refresh after being closed. See issue #173.
The steps to recreate:
- refresh the page
- zoom the first graph
- close the zoom
- add 'detached houses' to graph
- zoom again
- Expected: zoomed graph has two lines
- Actual: zoomed graph has only one line.

The basic issue was that the graph is a child of the Element el-dialog,
which is not destroyed when it closes. So the children don't get the
normal Vue lifecycle events.

Fix is to create a datum that can change in the parent, which will then
be noticed by a watch() in the child graph. This is the timestamp datum.