Additional fixes for #191, since the data is also processed by JavaScript on the client, and needs to cope with a variable number of results.
We make the assumption that missing values could in principle occur
anywhere in the series, so we normalise the comparison table to use the
maximal set of dates for any of the comparison series. We then ensure
that every series has some minimal value for every point in the series.