Add a list of exceptional locations
In preparation for #203, add a mechanism to the rake script which
loads the regions data into files to allow for the addition of
exceptional locations which are not in the data yet.

Specifically,

    data/location-exceptions.yml

is a YAML file of locations which can be recognised by the locations
finding tool, but which are not yet in the underlying data.

The intention is that these locations are not part of the normal search
workflow, but will bring up an exception warning message to the user,
e.g. suggesting that statistics for a particular location will be
available at some future time.