Detect and display forthcoming locations
For #203: if the user enters the name of a local authority that is yet
to be included in the statistics published by ONS, they will now get a
warning.

This fix is still provisional, as we're waiting for confirmation from
the product owner of the definitive list of location names, and the copy
for the warning message.