Major version upgrade of webpack
We had problems building the latest version of the app in production.
Simultaneously, we were several major versions behind with the webpacker
configuration, making it hard to follow on-line answers and
suggestions. The solution to both issues, I hope, was to rip-out all of
the webpacker configuration, and re-generate it with the webpacker rake
tasks.

This is now done, but I have low confidence that there are not some
lingering gotchas from this fairly major surgery.