Fix for "circular structure in JSON" issue
This commit fixes issue #15. The diagnosis was that Vue render function
has changed, so that if we pass a DOM element into the string renderer,
Vue will attempt to call `JSON.stringify` on some branches of the
render pathway. If the DOM element has been somehow made into a Vue
reactive value, it will have a reference to the Vue object, which has a
reference to the router, which in turn has a reference to the Vue
object. Hence the circularity.

In our case, we were rendering cells in the data table that had no value
as `span` elements with a class that allows the text to be italicised.
Replacing this with a simple (non italic) non-element string "no data"
prevents the problem from arising, at the cost of losing the italic
font rendering.