WIP commit, working on the regions data update
As part of the change for #217, I have updated the regions data file and
Alex has re-generated the geoJSON files from the ONS boundaries data. At
the moment, the keys in the generated geoJSON are not the keys we're
looking for the code, so map navigation is not working with the new data
yet.