Fix metadata field names from ONS
As part of the process for #217, we need to take the GeoJSON file that
was created from the ONS boundaries shapefile and fix the metadata.
There's no standard for the field names that ONS uses to denote GSS
codes, so they tend to change each year. This script looks for things
that look like GSS codes, then systematically rename them to a single
JSON property: `code` for the actual code, and `name` for the name of
the location.

The intent is that this script will be robust to use next time we have
to go through this process, but without seeing the actual data it's hard
to tell!