Add the skeleton mechanism for i18n in Vue
Adds the key pieces of `vue-i18n` to the use of Vue for runtime
internationalization. The messages are taken from the existing Rails
message catalogue files (`./config/locales`), which required adding a
YAML loader and configuring webpack to use it.

As proof-of-concept, we translate one table label in a single-location
view.

Issue #233