Pass the current locale from Rails
To ensure that we have consistent locales being applied in both Rails
and Vue, we pass the current locale from Rails via the `window.ukhpi`
state object. This is more reliable than checking URL arguments, since
we also detect if the user is selecting a language using the HTTP
`Accept-Language` header.