Simplify access to current locale in Vue components
We frequently need to refer to the current locale in Vue components,
e.g. to select the appropriate label for an object. I was adding a
computed property to each component, but this is not necessary as the
locale does not change during the lifetime of the Vue application (if
the user changes language, Rails will re-render the page first).

A better solution was to provide a `$locale` property globally on every
Vue component, by adding a property to the Vue object prototype.