Revert the default relative_url_root
As part of the configuration changes in readiness for building a Docker
container, I reset the default value of `R.a.config.relative_url_root`
to `''`. This commit resets it to `'/app/ukhpi'`, as that is more
consistent with the current deployment pattern via Chef. We can switch
to a Docker-friendly default in future