Relocate UKHPI documents into UKHPI project
Prior to this commit, the UKHPI documentation was stored in the
`lr_landing` project. With this change, we are moving the about page,
the pdf user manual and the DSD into the UKHPI project. The primary
reason for doing this is to enable us to also serve Welsh language
versions of some of those assets, particularly the about page and the
user manual. At the moment, the Welsh about page is a placeholder.