Fix missing LAs on map in Welsh lang mode
This commit addresses #248, in which local authorities were not shown on
the locations map when in Welsh mode. The base issue was that some LAs,
including some in Wales, did not have a `:cy` language-tagged name. The
fix I've adopted here is to ensure that in the pre-compiled
`locations-data.js` and `locations_table.rb` files, we maintain the
invariant that every location has two labels (`:en` and `:cy`), by
re-using the English language label if no Welsh language label was
given.

This fix is appopriate for some Welsh LA's, whose name is correctly the
same in both langauges (e.g. Gwynedd). It is more of an approximation to
say that the Welsh name for Mendip is also Mendip, but doing so keeps th
code simpler.

The rake task that compiles `locations-data.js` and `locations_table.rb`
will only add the faux-Welsh labels if an actual label is not present in
the source data. So if in future more Welsh labels are added to the
data, it will suffice to re-run the `rails ukhpi:locations` task.