WIP working on support for Welsh grammar muations
Preliminary commit of support for Welsh mutations. In certain
circumstances, we need to change words in constructed Welsh strings
depending on various patterns of consonants. We don't need to solve the
general problem, just enough to cover the cases that come up in
creating date and location templated-strings in UKHPI