Trap focus in modal dialogues
As contribution to #254, this commit adds a wrapper element that will
trap the focus within a displayed modal dialogue. Prior to this change,
a user of keyboard navigation would tab from the last active element in
the dialog to the next focusable element in the background application.
With the focus trap enabled, tabbing from the last active element will
return focus to the first active element of the dialog, and conversely
shift-tab from the first active element will cycle back to the last
active element in the form

Also logged as ticket #250