Add support for Welsh mutations
This commit addresses feedback from the Welsh translation changes in
issue #233. Specifically, we need to apply Welsh consonant mutations in
some cases. The mechanism this commit adds is possibly more than is
needed for the presenting issue, but we are expecting more mutions to be
needed in future.

This change addresses the feedback that some months should be mutated.
E.g:

    O Mai

should become:

    O Fai