Put Welsh mutations into practice
This commit identifies some of the places where we need to apply Welsh
mutations, and applies them.

I found it was necessary to duplicate the mutation machinery in
JavaScript for mutations of interactively selected locations and dates,
but this only requires a subset of the mutation rules (and, in
particular, we can assume proper-noun capitalisation).