Resolve issues with Welsh location names
This commit addresses the issue of missing Welsh location names in a
more nuanced way. #233

The core problem is that some locations do not have a name in Welsh in
the data (i.e. a lang tag of 'cy'). This may be because:

- the Welsh name and the English name are the same, e.g Gwynedd
- the location is in England, and there is not a standard Welsh name
  for that location. E.g Blackpool

The consequence of the missing Welsh names earlier in the project was
that those local authorities did not render on the map when in Welsh
language mode. This is undesirable for England, Scotland and NI, but
especialy problematic for Wales.

My first fix was to have the language table generator Rake task just
copy the English names to the Welsh name, where it would otherwise have
been missing. This addresses the map issue, but leads to a further
problem: in Welsh, we should mutate the consonants in some situations,
e.g. after prepositions 'o' and 'i'.  The copying fix removes the
distinction between English locations that do not have Welsh names and
those that do, meaning that we can't perform the mutation correctly. A
symptom of this problem is that "yn Blackpool" was mutated to
"ym Mlackpool". This is not correct.

My solution has been to adapt the Rake task to only copy the :en label
to the :cy label *if* the location is in Wales. The consequence of this
is that locations in England, Scotland and NI *may* not have a Welsh
language label. This is a major change of the invariant, because we can
no longer rely on, e.g. `labels[$locale]` to have a value. It does,
however mean that we can distinguish between non-Welsh locations that
have Welsh names (should be mutated) and non-Welsh locations that rely
on re-using the English name (no mutation).

While fixing these issues, I also noticed and addressed a couple of
others:

- missing GSS code for Northern Ireland
- redundant `features.json` file
- incorrect translations in print view

For the sake of simplicity, all of these are addressed in this commit.