Don't mutate the single-letter 'ch'
epimorphics/hmlr-linked-data#14
Our Welsh translator reported that the mutation rules were incorrectly
mutating 'i Chwefror' to 'i Ghwefror'. There **is** a mutation for 'c'
following 'i', but this does not apply to 'ch' which is a single letter
in Welsh.

I have updated the mutation rules to only apply in the case that 'c' is
not followed immediately by 'h'.