Fix 500 error when rendering not-understood resp.
epimorphics/hmlr-linked-data#15
This PR fixes a recently identified problem in UKHPI: if the user
changes the URL parameters so that the location is not recognised, it
should return a 400 error. Instead, the app was failing because there
was no `user_selections` object from which to determine the current
language.

This fix:

- adds the `@view_state` to the error page renderer, so that language
  selection will work correctly
- also lists URL validation errors that have been identified