Update README with ssh tunnel instructions
Access to the data API on the new infrastructure is different than it
was on the old system. After some help from @mika018 (thanks Mihajlo!) I
was able to get an ssh tunnel working correctly.

These instructions work on Linux (Ubuntu 21.04). They may need amending
and tweaking to cover MacOS and Windows.