Add Prometheus metrics for UKHPI
This commit adds the standard Prometheus metrics that are built in to
the Ruby Prometheus client, and adds API and memory usage, per local
best practice.

This version uses the `prometheus-client` 4.0 release, which includes an
metrics registry and helper methods, so we can dispense with our own
local Prometheus registry.