refactor: added `NoMethodError` clause
Also improves logging by using blocks instead of strings as Ruby has to evaluate these strings, which includes instantiating the somewhat heavy String object and interpolating the variables, and which takes time. Therefore, it's recommended to pass blocks to the logger methods, as these are only evaluated if the output level is the same or included in the allowed level (i.e. lazy loading). (Documentation)[http://guides.rubyonrails.org/debugging_rails_applications.html#impact-of-logs-on-performance]